requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Pillow==10.1.0
qrcode==7.4.2
reportlab==4.0.7
python-dateutil==2.8.2
Werkzeug==3.0.1
